+++ src/config.rs
@@ -382,13 +382,22 @@ impl ServerConfig {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(rename_all = "snake_case", deny_unknown_fields)]
 pub struct TestConfig {
-    /// Number of test executions to run in parallel. The default is `50`.
+    /// Number of test executions to run in parallel.
     pub parallelism: usize,
+    /// Delay between submitting initial test executions, in milliseconds.
+    ///
+    /// Once the `parallelism`` permits are exhausted, this throttle delay is
+    /// ignored and new tests are submitted eagerly as prior tests complete and
+    /// free permits.
+    pub throttle: u64,
 }
 
 impl Default for TestConfig {
     fn default() -> Self {
-        Self { parallelism: 50 }
+        Self {
+            parallelism: 50,
+            throttle: 100,
+        }
     }
 }